+++ apps/site/src/app/admin/applicants/ApplicantsSummary.tsx
@@ -1,0 +1,3 @@
+export default function ApplicantsSummary() {
+	return <p>PLACEHOLDER: Various summary charts</p>;
+}
