+++ apps/site/src/app/admin/applicants/page.tsx
@@ -1,1 +1,1 @@
-export { default as default } from "./Applicants";
+export { default as default } from "./ApplicantsSummary";
